DocTR Pipeline/requirements.txt
python-opencv
matplotlib
python-doctr[torch]
pycairo
python-doctr[torch]@git+https://github.com/mindee/doctr.git